Change the tweetcarousel component name, setup the ternary operations on the tweet components
